requirements.txt
Babel==2.9.1
bidict==0.21.4
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.4
controller==0.1.0
gTTS==2.2.3
idna==3.3
Jinja2==3.0.3
key==0.4
MarkupSafe==2.0.1
mysql-connector==2.2.9
mysql-connector-python==8.0.28
numpy==1.22.1
pandas==1.3.5
Pillow==9.0.1
protobuf==3.19.4
pynput==1.7.6
python-dateutil==2.8.2
python-engineio==4.3.1
python-socketio==5.5.2
pytz==2021.3
requests==2.27.1
six==1.16.0
tkcalendar==1.6.1
typing_extensions==4.1.1
urllib3==1.26.8
